examples/python/adk-demo/pyproject.toml
[project]
name = "adk-x402-demo"
version = "0.1.0"
description = "ADK x402 Payment Protocol Demo using refactored a2a_x402 package"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk",
    "google-adk>=0.5.0",
    "x402>=0.1.4",
    "click>=8.2.0",
    "eth-account>=0.13.7",
    "httpx>=0.28.1",
    "uvicorn>=0.34.2",
    "web3>=7.10.0",
    "cdp-sdk>=1.20.0",
    "a2a-x402",
    "docker>=7.0.0",
    "requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["server", "client"]

[project.scripts]
server = "server.__main__:main"

[tool.uv.workspace]
members = ["client"]

[tool.uv.sources]
a2a-x402 = { path = "../../../python/a2a_x402", editable = true }
